mainPyVR/requirements.txt
beautifulsoup4==4.8.0
Django==2.2.4
django-bootstrap4==1.0.1
pytz==2019.2
soupsieve==1.9.4
sqlparse==0.3.0
gevent
gunicorn
xlwt
djangorestframework
django-cors-headers